requirements.txt
pyinstaller
customtkinter
pyserial
requests